requirements.txt
Django==2.2.11
django-environ==0.4.5
django-extensions==2.2.8
django-taggit==1.2.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
